renaming temporal_network module
