+++ src/flowstab/temporal_network.py
