+++ tests/test_temporal_network.py
@@ -10,7 +10,7 @@ import pandas as pd
 from scipy.sparse import csr_matrix
 
 
-from flowstab.TemporalNetwork import ContTempNetwork
+from flowstab.temporal_network import ContTempNetwork
 
 class TestContTempNetwork:
     def setup_method(self):
@@ -151,77 +151,77 @@ def test_ContTempNetworkErrors():
 def test_ContTempInstNetwork():
     """
     """
-    from flowstab.TemporalNetwork import ContTempInstNetwork
+    from flowstab.temporal_network import ContTempInstNetwork
     pass
 
 def test_lin_approx_trans_matrix():
     """
     """
-    from flowstab.TemporalNetwork import lin_approx_trans_matrix
+    from flowstab.temporal_network import lin_approx_trans_matrix
     pass
 
 def test_compute_stationary_transition():
     """
     """
-    from flowstab.TemporalNetwork import compute_stationary_transition
+    from flowstab.temporal_network import compute_stationary_transition
     pass
 
 def test_compute_subspace_expm():
     """
     """
-    from flowstab.TemporalNetwork import compute_subspace_expm
+    from flowstab.temporal_network import compute_subspace_expm
     pass
 
 def test_csc_row_normalize():
     """
     """
-    from flowstab.TemporalNetwork import csc_row_normalize
+    from flowstab.temporal_network import csc_row_normalize
     pass
 
 def test_find_spectral_gap():
     """
     """
-    from flowstab.TemporalNetwork import find_spectral_gap
+    from flowstab.temporal_network import find_spectral_gap
     pass
 
 def test_remove_nnz_rowcol():
     """
     """
-    from flowstab.TemporalNetwork import remove_nnz_rowcol
+    from flowstab.temporal_network import remove_nnz_rowcol
     pass
 
 def test_rebuild_nnz_rowcol():
     """
     """
-    from flowstab.TemporalNetwork import numpy_rebuild_nnz_rowcol
+    from flowstab.temporal_network import numpy_rebuild_nnz_rowcol
     pass
 
 def test_sparse_lapl_expm():
     """
     """
-    from flowstab.TemporalNetwork import sparse_lapl_expm
+    from flowstab.temporal_network import sparse_lapl_expm
     pass
 
 def test_sparse_lin_approx():
     """
     """
-    from flowstab.TemporalNetwork import sparse_lin_approx
+    from flowstab.temporal_network import sparse_lin_approx
     pass
 
 def test_sparse_stationary_trans():
     """
     """
-    from flowstab.TemporalNetwork import sparse_stationary_trans
+    from flowstab.temporal_network import sparse_stationary_trans
     pass
 
 def test_set_to_ones():
     """
     """
-    from flowstab.TemporalNetwork import set_to_ones
+    from flowstab.temporal_network import set_to_ones
     pass
 
 def test_set_to_zeroes():
     """
     """
-    from flowstab.TemporalNetwork import set_to_zeroes
+    from flowstab.temporal_network import set_to_zeroes
     pass
